[mod] cleansing genre: overhaul
I've been working on cleansing the Genre collection / class. It was one of the simpler API data collections, so it makes the most sense to do first.

Despite the fact that it small, I've noticed a lot of things that need to be cleaned up — not only in this collection, but in all collections. It may be worth it to create a base class, and extend all other collections of of it, because they will all share a lot of functionalities. Examples: `.matches( )`, `.combine( )`, and even `constructor`. More research will be needed.

Nonetheless, I feel satisfied with where the Genre class has landed, and it's a good start to what will come.